feat(tests): total-time speed check, labeled settings, fractional advisory time
- Speed check now judges the session as a whole (minTimePerQuestion x question
  count) against the server clock instead of flagging any single fast answer;
  the result dialog names the required total.
- Settings fields in the test editor get small headings above each control.
- Advisory time accepts fractional minutes (Int -> Float, migration included).
- Attention check and trap duplicate are mutually exclusive in the question
  dialog and rejected together by the API.

+++ packages/shared/src/validation/test.ts
@@ -18,7 +18,8 @@ export const updateTestSchema = z.object({
   title: z.string().min(1).optional(),
   description: z.string().nullable().optional(),
   responseCap: z.number().int().positive().nullable().optional(),
-  advisoryTimeMin: z.number().int().positive().nullable().optional(),
+  // Fractional minutes are allowed (0.5 = 30 seconds), so this is not an integer.
+  advisoryTimeMin: z.number().positive().nullable().optional(),
   minTimePerQuestion: z.number().int().min(0).optional(),
   demographicFilters: demographicFiltersSchema.nullable().optional(),
 });
